requirements.txt
python-telegram-bot
requests
schedule
openai
python-dotenv
craiyon.py
feedparser
